+++ backend/application/__init__.py
@@ -15,11 +15,8 @@ from marshmallow import ValidationError
 from application.extensions import db, cors, jwt
 from .routes import auth_route, auth_route
 
-from flask_jwt_extended import JWTManager
-from flask_cors import CORS
 
 def init_app(config):
-
     app = Flask(__name__)
     app.config.from_object(config)
     db.init_app(app)
