+++ backend/application/routes/auth_route.py
@@ -1,9 +1,7 @@
-from flask import current_app as app
-from flask import jsonify
-from flask import request
-from flask import Response
+from flask import jsonify, request, abort
 from . import api
-from functools import wraps
+from application.models.staff import Staff
+from application.extensions import db
 
 from flask_jwt_extended import (
     jwt_required,
@@ -13,44 +11,34 @@ from flask_jwt_extended import (
     set_access_cookies,
     set_refresh_cookies,
     unset_jwt_cookies,
-    verify_jwt_in_request,
-    get_jwt,
 )
-
-
-def admin_required():
-    def wrapper(fn):
-        @wraps(fn)
-        def decorator(*args, **kwargs):
-            verify_jwt_in_request()
-            claims = get_jwt()
-            if claims["admin"]:
-                return fn(*args, **kwargs)
-            else:
-                return jsonify(msg="Admins only!"), 403
-
-        return decorator
-
-    return wrapper
 
 
 @api.route("/login", methods=["POST"])
 def login():
     email = request.json.get("email", None)
-    password = request.json.get("password", None)
+    # password = request.json.get("password", None)
 
-    if email != "test" or password != "test":
-        return jsonify({"msg": "Bad email or password"}), 401
+    results = (
+        db.session.execute(db.select(Staff).where(Staff.email == email)).scalars().first()
+    )
 
-    # Create the tokens we will be sending back to the user
+    if results is None:
+        abort(404, description=f"Staff with {email} not found.")
+
+    data = results.json()
+
+    if data["email"] != email:
+        return jsonify({"msg": "user is not registered in system"}), 401
+
+    # Create the tokens
     access_token = create_access_token(
-        identity=email, additional_claims={"admin": False}
+        identity=email, additional_claims={data["access_control"]: True}
     )
     refresh_token = create_refresh_token(identity=email)
 
-    # Set the JWTs and the CSRF double submit protection cookies
-    # in this response
-    resp = jsonify({"access_token": access_token, "refresh_token": refresh_token})
+    # Set the JWTs and the CSRF double submit protection cookies in this response
+    resp = jsonify({"status": "success"})
     set_access_cookies(resp, access_token)
     set_refresh_cookies(resp, refresh_token)
     return resp, 200
@@ -59,12 +47,8 @@ def login():
 @api.route("/token/refresh", methods=["POST"])
 @jwt_required(refresh=True)
 def refresh():
-    # Create the new access token
     current_user = get_jwt_identity()
     access_token = create_access_token(identity=current_user)
-
-    # Set the access JWT and CSRF double submit protection cookies
-    # in this response
     resp = jsonify({"refresh": True})
     set_access_cookies(resp, access_token)
     return resp, 200
@@ -76,15 +60,10 @@ def logout():
     unset_jwt_cookies(resp)
     return resp, 200
 
-
-@api.route("/example", methods=["GET"])
-@jwt_required()
-@admin_required()
-def protected():
-    app.logger.info("Protected endpoint accessed")
-    email = get_jwt_identity()
-    # resp = Response(jsonify({"hello": "from {}".format(email)}), 200)
-    # resp.headers["Access-Control-Allow-Origin"] = "http://127.0.0.1:5173/"
-    # resp.headers["Access-Control-Allow-Credentials"] = "true"
-    # return resp, 200
-    return jsonify({"hello": "from {}".format(email)}), 200
+# Endpoint used to test protected routes
+# @api.route("/example", methods=["GET"])
+# @jwt_required()
+# @admin_required()
+# def protected():
+#     email = get_jwt_identity()
+#     return jsonify({"hello": "from {}".format(email)}), 200
