+++ backend/application/routes/route_decorators.py
@@ -1,0 +1,24 @@
+from flask import jsonify
+from functools import wraps
+from application.enums import AccessControlRole
+
+from flask_jwt_extended import (
+    verify_jwt_in_request,
+    get_jwt,
+)
+
+
+def admin_required():
+    def wrapper(fn):
+        @wraps(fn)
+        def decorator(*args, **kwargs):
+            verify_jwt_in_request()
+            claims = get_jwt()
+            if claims[AccessControlRole.ADMIN.value]:
+                return fn(*args, **kwargs)
+            else:
+                return jsonify(msg="Admin Rights Required"), 402
+
+        return decorator
+
+    return wrapper
